requirements.txt
Django==4.2.26
django-allauth==0.51.0
django-otp==1.5.0
qrcode==7.3.1
Pillow==10.0.1
requests==2.31.0
openai==1.3.6